# Optimise `mrb_iv_get`